+++ lua/md-viewer/commands.lua
@@ -37,6 +37,7 @@ function M.setup()
   vim.api.nvim_create_user_command("MdViewerFindPrevious", function() controller.find_previous() end, {})
   vim.api.nvim_create_user_command("MdViewerBack", function() controller.history_back() end, {})
   vim.api.nvim_create_user_command("MdViewerForward", function() controller.history_forward() end, {})
+  vim.api.nvim_create_user_command("MdViewerToggleLineMarkers", function() controller.toggle_line_markers() end, {})
 end
 
 return M